+++ pulledpork.py
@@ -35,7 +35,7 @@ from tempfile import gettempdir             # temp directory mgmt
 from time import strftime, localtime
 from urllib.parse import urlsplit           # get filename from url
 
-from pp3 import logger                      # Our logging helper (for now)
+from ppork3 import logger                   # Our logging helper (for now)
 
 
 # -----------------------------------------------------------------------------
